Set postgresql search_path via ansible
Set here so that the sql scripts find all of the tables
and functions in the various schemas. Generated must
be the first schema.

Set after a flush_handlers because the configuration changes
made in the azavea.postgresql role have not yet been applied
in the case of a new provision.