Tweak OSM download to pull all data within bbox
The analysis uses OSM data other than highways, such as
amenity=school. For now, while in development, assume we need
all the data within the bounding box. We can tighten up these
parameters once the analysis finalizes and we deploy the
application.

As a point of reference, the center city philly bbox pulled
15mb of data in ~60s. According to the Overpass wiki page,
we can continue to make requests to the API as long as we don't
exceed these limits:
- < 10,000 queries per day
- < 5GB download per day

http://wiki.openstreetmap.org/wiki/Overpass_API

If we do start to get near these limits, we'll need to
investigate other options for download.