Don't map `/srv/dist` as a volume in nginx
Mapping `/srv/dist` in the nginx container to the host directory only
works in development and would have to be overridden in production. And
it's slightly non-trivial to override docker-compose config parameters
with nothing. So this just removes the volume and changes
`scripts/update` slightly so that the files will be copied in before
the nginx container is built, like they are in cibuild.