Upload analysis results to S3
Makes `export_connectivity.sh` upload the results (in a zip file,
since they deflate quite a bit) to S3 if the AWS_STORAGE_BUCKET_NAME
variable is set.

For it to work in development, the AWS_PROFILE variable will also need
to be set and the `-v /home/vagrant/.aws:/root/.aws` will need to be
added to the `docker run` command.

Also makes "output file prefix" into a required argument for
export_connectivity.sh.