Download and unzip boundary shapefile
Makes entrypoint.sh look for `PFB_SHPFILE_URL` and, if set, download and
unzip the file, then set the PFB_SHPFILE to be the unzipped .shp file.

Also makes a temporary directory to put the file in and sets
NB_OUTPUT_DIR to that if it hasn't been set explicitly already.