Replace NB_BOUNDARY_BUFFER with NB_MAX_TRIP_DISTANCE
Connects #81.

Replaces the NB_BOUNDARY_BUFFER variable used in the analysis with
NB_MAX_TRIP_DISTANCE, since that's what it's actually doing.

In import_neighborhood.sh, where the buffer was used to truncate the
census blocks table to only the ones we care about, it now uses half
of NB_MAX_TRIP_DISTANCE, since it's not possible for a trip between
two points inside the boundary to go more than NB_MAX_TRIP_DISTANCE/2
outside of it.

NB_BOUNDARY_BUFFER is still accepted as an override to the census block
filter, in case that should prove desirable.

Also includes a few fixes and tweaks to `run-local-analysis`