Set default output SRID in scripts to 2163
It's a bit academic because we don't expect to be calling these scripts
without setting the input SRID, but since there's a default defined it
might as well make sense. Since nb_max_trip_distance expects meters,
setting the default to the U.S. National Atlas projection, which is in
meters, seems as good a choice as any.