Change usage message of streetlight_gates
Changes the SRID and default buffer in the usage message, but *doesn't*
change this script to use nb_max_trip_distance since this seems to
actually be used as an actual buffer in this query.
The script currently isn't pulled into the analysis anywhere, but when
it is we'll need to make sure it's supplied with that parameter.