Convert AnalysisJob status to a @property
Turns the property into a @property that's derived from the latest
AnalysisJobStatusUpdate and, when applicable, from batch job status.
Adds 'QUEUED' to cover the period between when a batch job is created
and when it starts sending status updates.

Don't show batch status as job status

Instead of passing through the batch status when jobs are in stages
before they get run and start sending their own updates, just use QUEUED
for all of them.