Update job cancel logic to handle inconsistencies
Before this commit, it was possible to cancel a job
but then submit further updates while the job continues
to run before being killed externally by AWS Batch,
leaving job status in an inconsistent state.

This attempts to improve this by providing a "safe" method
to update job status with by only submitting further updates
if the job is not cancelled.

In addition, the job.cancel() logic was tweaked to attempt to
better prevent the race condition above by cancelling the job
first, then making the cancel request to Batch. If the Batch
request fails, the cancel state is reverted to allow later
attempts to cancel. Unfortunately this also creates a race
condition in the other direction, where a job update could
be dropped while the job is in a cancelled state, but this
seems less of an issue than the other case.