Provide results path to analysis and tiling tasks
Rather than hard-coding the path setup on S3 into the scripts run by the
analysis and tiling tasks, define them on the AnalysisJob model and pass
them in as environment variables to the tasks.

Also changes the ogr2ogr command to overwrite, which won't be relevant in
production but is convenient in development with a mapped /data/
directory.