Add overall_score annotation to job endpoint queryset
The overall score is in the jsonb field, but that makes it not super easy
to get at and, I think, impossible to sort on with OrderingFilter.
Since we want an 'overall_score' property in the results and want to be
able to sort by it, adding it as an annotation to all jobs makes sense.