Keep successful jobs as latest until newer jobs finish
To avoid the situation whereby starting new analysis job would make the
scores for a neighborhood disappear from the default list (which will be
filtered to latest), changes the 'latest' filter to take the latest
successful job for a neighborhood if there is one, falling back to the
last modified if not.

Converts the filter from using a boolean annotation to doing the
filtering in code and using a `pk__in` query, which is less efficient
but much easier to implement.  The query expression to do all that
conditional filtering and aggregating would likely be formidable.