Make tile generation command less quiet
I had added the --quiet flag to the tile generation command because its
non-quiet behavior is to print a line for every tile (so 165,235 lines
in the case of NYC).  But since it takes a while, no output is not very
nice either because it's hard to tell if it's working. So this removes
--quiet but pipes it to an awk command that filters the output down to
one in every 1000 lines.