Closes #273: Save job names/defs to AnalysisJob model
Saved via lazily set database field that shadows the original
property. Unable to use ModelField(default=callback) option
because we need the job uuid to create the job name.

Other option was to make job name fields DB fields instead of
properties directly, but that requires additional save logic
and a data migration that would take more time to write,
and we don't filter/sort on these fields anyways. Could
still migrate to that later if necessary, as this puts the
first half of that machinery in place.