Include neighborhood object in job api results
Switches the analysis-jobs endpoint from including UUID and label as
properties on the job to including `neighborhood` as an object, with
those details plus state and point geometry inside.

Since DRF doesn't have easy automatic behavior for saving inlined related
objects, I borrowed an idea from MUB and made neighborhood on the
AnalysisJob serializer into a field that expects UUID for create/save
operations but includes the serialized object when serializing for
presentation.