Separate navbar and page/view logic
A couple views (job list and user list) were using the `isAdminUser`
property of the navbar controller to decide whether to show their 'Add'
buttons or not.
Which was fine until I added a `scope` to the navbar directive, which
caused the DOM to be built differently so that the navbar controller was
a peer rather than a parent of the list views.

Rather than fix the dependency, I removed it, having the list views go
right to the source (AuthService) for that information.

In the other direction, the scope variables I had added were 'admin' and
'showTitle', which enabled views to influence how the navbar displayed.
So that introduced a link in the other direction (pages->navbar) that was
also not idea.  In the course of trying to fix the first issue, I also
refactored the navbar controller to get the information it needs from
`$state` rather than from scope parameters.

To properly determine whether a view is an 'admin' view, I made an
'admin' abstract route and put all the views with `/admin/` URLs into it.

So this touches a lot of files, but what it's doing is fairly simple and
I think it leaves things a bit cleaner.