Add census blocks raster tile generation
Styled by 'OVERALL_SC' with a 10-step color ramp from red to blue via
purple. Went with 70% opacity because it seemed pretty vivid and still
afforded some view of the base map.

I tried it with border styling, but the block borders are pretty much
roads, so drawing them mostly just obscures more interesting things.

As with neighborhood_ways, uses the shapefile because it's way smaller
than the geojson, so quicker to download, and I don't know of any
disadvantage of doing it that way.

Converts pretty much all of `run_tilemaker.sh` into a function so that it
can be called once for each export file to be tiled.

Remove census block borders and decrease opacity

The borders are pretty much roads, so drawing them mostly makes clutter
and obscures more interesting things. Also made the fill a little more
transparent.  It's plenty vivid still, so might as well be able to read
the base map.

Actually-working styles for census_blocks