Mount fewer volumes in tilemaker and analysis
The volumes in common.yml are the ones the django container wants, but
some of them don't make sense for the analysis and tilemaker containers,
and `data` is not a good idea in the case of tilemaker because it writes
its working version of shapefiles there, possibly causing contention or
confusion.  This moves the volumes out of common and sets them on the
containers in a more fine-grained way.