Revert "Attempt a single retry in case of job failure"
This reverts commit 9132936f9bdef78103f45226f772370c36cc2725.

The automatically installed version of botocore we're currently
using isn't new enough to support the changes in the Batch API
which support retryStrategy. Not willing to mess with
dependencies ATM without a review cycle, so this goes away for
now.

Note also that in testing the second retry attempt failed
consistently attempting to start the database process,
so the addition of retryStrategy holds no value without
further debugging.