Add generated tiles to layer picker
Adds layers for the generated tiles to the layer picker. Also removes
the base map from the picker, so it's just always loaded, and moves the
boundary from the overlays group to the base layers group, since it
should be available in combination with the feature layers but wouldn't
work well in combination with the tile layers.