Add None option to Destinations layer picker
The fix for this is awkward. In order to not mess up the
order in which the layer groups draw in the picker, we
have to ensure that we add the "None" layer immediately
before the remaining destinations layers. The group layer
control doesn't give us a way to specify a particular order.

If we only add based on whether the ctl.destinationsNoneLayer
exists or not, we create it too early during the first scope
resolution, and the Destinations draw before the Overlays.

If we just create based on whether layers are present in the
$q.all callback, we may double create the none layer.

Opened:
https://github.com/ismyrnow/leaflet-groupedlayercontrol/issues/54