Release OSM cache upload lock on error
Delete the lockfile from the bucket regardless of what happens with the
upload.  This means we could end up with a series of jobs trying and
failing in the same way, but that's preferable to allowing a single
failure to leave things in a state where human intervention is required
to clean up and resume processing.