Delete lockfile if an error occurs while getting lock
It's unlikely but not impossible that something could cause an error
after a job has written the lockfile to download an OSM extract but
before it has finished checking whether it has the lock. In that case
it could leave an orphaned lockfile. This adds a catch-and-raise
(`finally` won't work because in non-error cases we don't want to delete)
to clean up. This could mean a job deletes a lockfile that was actually
written by another, but the failure mode of that is acceptable.