Switch the 'waiting on lockfile' message to INFO level
When running two jobs for the same state at the same time, it's not clear
from the logs that the second is waiting for the OSM extract lockfile
created by the first to disappear before it tries to download the extract
from S3.  It's possible to infer that from the timing, but it's worth
logging it explicitly all the time, not just in debug mode.