Add Tilegarden to docker-compose
Adds the tilegarden container to the docker-compose config. It wasn't
immediately obvious how to change claudia-local-api to use a different
port, so I just mapped the default port inside the container to a chosen
one outside.

Also splits the "common" docker-compose config, because I wanted to
inherit the environment variables but it seemed too weird for this node
container to inherit a "django" config.