Stop transpiling tilegarden code
Since this uses a Node 8.10 runtime, most newer EMCAScript features should
be supported without transpilation. It appears that the only thing missing
that we're using is the import/export syntax. So this refactors our imports
to use `require` and `module.exports` and removes all the babel pieces.

Since we're not using a 'bin/' directory anymore, this also changes the
config compilation command to put the xml back in the source directory
alongside the mml and mss. And adds the resulting files to .gitignore.

This also adds `ncc`, which will hopefully work for making an efficient
deployment bundle, but that part is not actually tested.  Putting it in
as a placeholder seemed better than either leaving the babel commands in
place or ripping out all the bundling-related pieces then having to add
them back later.

Add compiled config files to .gitignore