Add Job ID to shapefile export
Include job id in shapefile import management
command and run command during analysis export.

This will slow down analysis jobs but not signficantly.

Jobs should run far enough apart that it doesn't lead
to excessive DB load during a large analysis run (but
this is untested).

While inefficient because we need to re-download the
shapefile we just uploaded, the shapefile sizes aren't
huge and it's cleaner than attempting to include
custom logic in bash to directly copy data from the
analysis DB to the app DB.