Scaffold Tilegarden Cloudfront CDN
Sketch out the interaction between Tilegarden and a CloudFront CDN
in deployment configs. In addition, adjust the TILEGARDEN_ROOT settings
variable in Django so that it corresponds to an environment variable
that can be propagated into Django containers.

Note that the infrastructure in this commit has not been tested on
staging yet, so there are likely further tweaks necessary in order
to get it fully operational.