Fix invisible admin pagination buttons
It turns out fixing a style syntax error in
https://github.com/azavea/pfb-network-connectivity/pull/655/files#diff-5645138bbcfd38020bf72c8203017f60L43
caused that style to start getting applied, but when combined with the
fact that the pagination buttons in the admin were getting the default
`color` style from `btn` (because they were using a non-existent
`btn-success` class), made them all but disappear.

This changes the class to `btn-default`, which is the same one used by
the pagination links on the public-facing site.

Resolves #674.