Read settings S3 bucket from querystring
Most of the code to read settings from S3 is still in, though the config
name is now a path parameter, but the part to provide the bucket, which
also toggles looking for S3 config in the first place, was dropped.
This is only really useful for debugging, but it doesn't do any harm.