Optimize Mapnik/PostGIS config and stop compiling config
This does a few related things:
- Adds some optimization parameters to the Datasource sections of the
layer config files.

- Since it doesn't seem like these options can be specified in the
CartoCSS/MML format, this means we have to switch to maintaining the
Mapnik XML by hand.  So the MML and MSS files are still in the repo, but
now they're just working files. The carto->mapnik config compilation
script is still there, too, but doesn't run automatically.

- Since the database parameters were being copied from the environment
into the config XML during compilation, that setup needed to change, too.
It's now handled at runtime, so the XML files now contain the environment
template bits that were formerly in the MML. This has the advantage that
it doesn't put extra copies of the secrets on the filesystem, and it also
prevents the issue whereby your local config would stop working after
building and deploying.

- To fill the database parameters in at runtime, they need to be set in
the Lambda function's environment, so this adds them.

- The 'deploy' commands in package.json were already crazy, and adding
more environment variables would have made them even worse. So this moves
them into scripts.