Add warming endpoint to Tilegarden
Since we want to be able to trigger Lambda warming based on API activity,
it needs to be an API endpoint rather than a scheduled task.

This implements that endpoint, which composes and launches a new invocation
of the Lambda (invoking the function directly, not going through API Gateway)
to trigger the 'lambda-warmer' package to do its thing.