Add Tilegarden-warming task
Adds a task to send a warming request to Tilegarden.  The sending and
timeout logic is on the tracking model (TilegardenWarmingEvent), so the
task just checks whether the last warming event is expired and makes a
new one if so.

The idea is to trigger this when someone first interacts with the site
so that by the time they get to the point of clicking on a neighborhood
the tiling lambdas will be ready to go. The analysis job list view is
called from the home page and the places list, so it should be sensitive
enough but not too noisy.