Ignore job ID in imported shapefiles
When importing a shapefile, the job ID should always be set to the job
we're importing it for.  We've hit a few different issues from missing
or mis-matched job IDs in input shapefiles, but we can just ignore the
value (or lack thereof) in the shapefile completely and set it to the
target job ID after import.

So this does that. Since there doesn't seem to be a way to set the value
of a field to a literal string via the layer mapping, this relies (as we
were doing before in the event of job id errors) on setting the job ID
after import on all geometries with no job ID.  Since this could
theoretically cause an issue if multiple imports were running
simultaneously, this puts the import in a transaction so there won't ever
be ID-less geometries actually committed to the database.