Define Tilegarden executor role in Terraform config
The Tilegarden Lambda function needs some permissions, specifically
- write logs
- attach to a VPC
- write to the S3 cache
- invoke itself for warming

If you don't provide a role to Claudia.js, it makes one that has the
first two, but since the other two are for custom Tilegarden functionality,
they need to be added.  The most straightforward way to do that is to define
the whole role in Terraform and just hand it to Claudia.js in the .env file.

So this adds the Terraform code to do that, and moves the `LAMBDA_ROLE`
variable out of the "Optional" section of the .env.example file.