Add documentation section and infra commands for migration
Adds a section to the deployment documentation about how to safely apply
migrations.  Since in some cases migrations need to be applied before
the API service is updated, also adds a pair of commands to `infra` to
deploy only the management task, so that migrations can be run before
the matching API service is updated and starts handling requests that
might rely on the schema being updated.