Add Terraform var for Tilegarden function name
Adds a tfvar for the name of the Tilegarden Lambda function and uses it
to restrict the Lambda execution permission on the TilegardenExecutor role.

See [discussion on PR #719](https://github.com/azavea/pfb-network-connectivity/pull/719#discussion_r260855692))

Also
- Uses the tfvar to set the name of the executor role.
- Renames the variable on the Tilegarden side to 'LAMBDA_FUNCTION_NAME',
since that's what it's used for.
- Removes the 'LAMBDA_ROLE' variable from the Tilegarden .env example and
uses the 'LAMBDA_FUNCTION_NAME' value with 'Executor' appended to match
what Terraform does when creating it.