Add Terraform config to create Tilegarden warming scheduled event
Adds a file to the Terraform config to define a CloudWatch event to trigger
`lambda-warmer` events on the Tilegarden function every 15 minutes.

Unfortunately this creates more circular dependency between Terraform and
Claudia--the permission to allow CloudWatch to invoke Lambda functions has
to be a policy on the targeted Lambda function, but when spinning up a
stack that Lambda function doesn't exist yet.  The problem can be solved
with dummy values and iterative bootstrapping, but it's too messy.

If this deployment ever changes so that the Lambda function is either
created by Terraform or always created before the Terraform config is
applied, this could be enabled.