Add subdivisions to 'countries' list, and country config
Main changes:
- Adds a config dictionary in settings to record what countries we want
  to track subdivisions for.
- Adds the `pycountry` package, which provides subdivisions
- Uses the `pycountry` and the config to add subdivisions to the list
  returned by the `countries` endpoint, only for the countries
  whitelisted in the config.

Incidental changes:
- Renames the country abbreviation field from `abbr` to `alpha_2`. That's
  what it's called in pycountry. Not that we have to be bound by that,
  but I like it better--more explicit.

This leaves the `state` endpoint alone. I think it should go away, but
that seems to fall under issue #732.

Though it adds `pycountry`, this doesn't remove `django-countries` or `us`.
It does reduce the usage of `us`, so I think we should be able to remove
that by the time this is all done.  Whether `django-countries` can/should
go away will depend on whether we want to keep using `CountryField`.