Convert Neighborhood.state_abbrev to CharField
Since we'll be supporting non-US states, the `state_abbrev` field on
Neighborhood can't be a USStateField anymore. So this makes it just a
CharField (longer than strictly necessary, but it doesn't do any harm to
have a little extra space in case some subdivisions have standard
abbreviations that are a little longer...

Since that was the only use of django-localflavor, this removes it from
the apps list, but doesn't remove it altogether because it's used in an
old migration.