Pull new base images on 'update'
The --pull option to `docker-compose build` makes it always look for a
new version of the base container. It does nothing if the image is up to
date, so it doesn't slow things down, and it should help prevent surprises
from image changes that break in CI where new environments get built often
but don't affect development because we tend to end up using cached images
for long periods.