Make next/prev buttons on Places more self-contained and DRY
The global params vars threw me off for a minute, so I started thinking
about how the next and previous buttons could be implemented without
them. So rather than `getNext` and `getPrev` being static functions that
use global variables, this switches to generating new functions with
the params baked in whenever the data changes.

It also seemed like an opportunity to reduce the number of pieces with
separate 'next' and 'previous' versions.