Stop using COUNTRY_CONFIG for label format
The label format as described by PFB is regular enough that just joining
state/province and country abbreviations will do what we need. And it's
way simpler than trying to encode it as a template in the config and pass
all possible template variables to that unknown template in the model
method.