Add countries to config and make subdivisions optional for some
Adds the countries identified in the document from PFB to the country
config and adds subdivision support for most of them.

This also makes "State/Province" (aka subdivision, aka 'state_abbrev')
optional in the Neighborhood creation form for some countries. This will
enable the use case of running the analysis for a whole region (e.g.
Gelderland, NL, can be entered with a label and country but no state, to
avoid having it show up as "Gelderland, GE, NL").

Denmark and Norway will need hand-made subdivisions lists. Denmark uses
numbers for region codes and 'pycountry' doesn't have the desired
subdivisions for Norway.