Clarify purpose of AWS_PROFILE in local analysis readme
The `run-local-analysis` script is set up so it will only cause results
to be uploaded to S3 when both `PFB_JOB_ID` and `AWS_PROFILE` are set.
But the documentation doesn't make that clear, so it seems like providing
an `AWS_PROFILE` should cause the files to be uploaded to S3 regardless
of whether there's a job ID. This hopefully clarifies the situation a bit.