Update Vagrant base box and Ansible version
Upgrades the Vagrant base box, since it was still on Trusty, which won't
be supported for much longer.

We don't rely all that much on the VM, since it's just used as a controlled
environment to build and run the containers, so upgrading is fairly low-impact.
But it turns out that the newer official Ubuntu base images have minimal
disk space--the old 14.04 one had 40GB, but their 16.04 box has 10GB--so
I switched to a Bento version of a 16.04 base box. (I tried a Bento 18.04
image, too, but it failed to build due to an issue with the apt repositories).

Upgrading to Python 3 means we can get rid of the extra steps that were
needed to get a version of the Python 2.7 SSL packages that Ansible would
accept.