Change _get_count function to get_count method
See https://www.django-rest-framework.org/community/release-notes/#380,
4th from the bottom:
"Add LimitOffsetPagination.get_count to allow method override #5846"

It's nice that DRF has a complete changelog, but it doesn't seem to be
possible to figure out what all the breaking changes for a project might
be from reading it. Though arguably in this case we wouldn't expect to
be notified since we were importing an internal function.