Don't specify ansible version in Vagrantfile
The base box comes with Ansible, and the version we're specifying doesn't
match, which causes provisioning to crash.
Just leaving off the version lets it use whatever comes with the VM,
which seems fine. Or at least better than making it significantly more
complex in order to pin the version.