Install osm2pgrouting and osm2pgsql with apt
The compilation of the specified version of osm2pgrouting started failing
because it or some dependency uses some outdated syntax.
There are versions of the that and osm2pgsql in the repo that are newer
than the ones we've been specifying. So I switched them to get installed
by apt with most of the other dependencies.
`quantile` could be installed using `pgxn`, but we don't have that
installed, so continuing to install it from source works fine.