Prevent Neighborhoods from getting deleted along with analysis jobs
The Neighborhood model has a foreign key to AnalysisJob to keep track of
the latest job for each neighborhood.  But the `on_delete` property of
that foreign key was set to `CASCADE`, so deleting a job would take the
neighborhood with it.  Which is not what we want.  This changes it to
`SET_NULL`, i.e. if the job that's currently the "last job" of a
neighborhood gets deleted, the `last_job` on that neighborhood gets
set to NULL and the neighborhood is fine.

Resolves #644