Restore ansible version specifier in Vagrantfile
When doing the Python upgrade I changed `ansible.install_mode = "pip"` to
`pip3` (a22eee6). But that's not actually a valid value of that variable (see
https://www.vagrantup.com/docs/provisioning/ansible_local.html#install_mode)
Which I guess made Vagrant silently ignore it and start installing Ansible
from apt. So once the Ansible version in the apt repository changed, that
caused provisioning to fail because `ansible.version` didn't match the
system-installed version. Not realizing it was installing Ansible the wrong
way, I resolved that issue by removing the version specifier (eb8d99ba).
But now that the version of Ansible in the repo has been upgraded to 2.9,
there's a new crash due to breaking changes to the ansible file spec.

So this sets `ansible.install_mode` back to `"pip"` and restores the version
specifier. Together, those succeed in installing Ansible 2.8.7, which works
with our roles.