Delete Neighborhood and AnalysisJob files from S3 on instance deletion
Now that we can delete neighborhoods--which means we can also delete analysis
jobs, since they get deleted along with their neighborhood--we run the risk
of creating a mess of orphaned files on S3.  The Neighborhood model keeps its
boundary file on S3, and the results of the AnalysisJob processing are stored
on S3 as well.
This adds post-delete handlers for those two models to delete the files from
S3 after the instances have been deleted.
For Neighborhood, which uses a `FileField`, it's just a matter of calling
`.delete()` on it.  AnalysisJob doesn't have a field to track its own assets,
but it does know the S3 path where they live, so we can use that to delete
them with boto3.