Add toast message for neighborhood deletion and adjust settings
PR #768, which enabled neighborhood deletion, should have included toastr
messages to convey the results. This adds those, and also adjusts the
toastr settings so the messages will show for slightly longer and be
positioned at the bottom right, rather than right on top of the header
text where they're hard to read.