Improve error messages for neighborhood and job operations
Fixes the bug where we were treating text as an array, and more broadly
cleans up and expands the handling of error messages from the API for
neighborhood and job interactions.  Since there are several possible
cases to handle and they're nearly the same between the two, this breaks
the parsing/formatting into a helper function, set up as a filter.

This also overrides the `save` method in the neighborhood serializer
so that "object already exists" errors can be caught and the message
sent in the response. Previously they were coming up as generic 500
errors.