Change neighborhood 'name' field on label change
The 'name' field on Neighborhood is a slug field derived from its label.
We have a uniqueness constraint on (name, country, state_abbrev, organization)
to prevent people from creating indistinguishable duplicates.  But the name
field was only being set if it wasn't already, which makes sense when you
can only create neighborhoods, but now that you can edit them it means you
could make two neighborhoods with different labels then edit them to be
identical.  So this just removes the condition so that we'll always derive
the 'name' when saving, and always raise the error if there's a conflict.