Use detailed geometry for neighborhood geometry preview
The simplified geometry returned by the 'bounds' endpoint is intended for
use in thumbnails.  When you put it in a bigger, zoomable map it looks
pretty rough.  So this makes the endpoint look for a 'detailed' query param,
and if it's there/truthy, return the full, unsimplified geometry. And it
sets that parameter for the request made by the neighborhood detail map
preview controller.