Save new geometry when editing neighborhood
The edit neighborhood form would accept a new geometry file but not
actually do anything with it, so I hid it in editing mode for PR #778.
This puts it back and makes it work, by modifying the check in the method
that `Neighborhood` uses to save geometry.  It was saving only if explicitly
told to override or if there was no geometry on the model.
Passing an explicit override would be tricky, since there are several layers
of DRF machinery between the request and the final save. So instead this
looks at the `boundary_file` field and uses its `_committed` property to
tell if the one currently on the model is old or new.