Make bounds endpoint return unsimplified geometry by default
We're using the `neighborhoods_bounds_geojson` endpoint three times--for
thumbnails, boundary overlay on the public site detail page, and boundary
preview on the neighborhood editing page.  The simplified geometry is only
worth having in the first of those--on the detail and edit pages, where
there's only one geometry being requested, the possible roughness of the
simplified boundary isn't worth the savings in transfer size.
So this changes the endpoint and service from
"defaults to simple, pass 'detailed' to get the detailed version" to
"defaults to detailed, pass 'simple'..."