Cache downloaded census data files to S3 within analysis
The analysis already looks on S3 for cached census data files (jobs and
blocks), but there's no facility for actually filling that cache.
This makes the analysis scripts that download and use those files also
push them up to the S3 cache so they'll be there for the next analysis run.
The files don't change (new versions would be released as new files, with
a newer year in the filenames), so we don't need to worry about rotating
or invalidating the cache.

Note: this protects the "upload to cache" steps from trying and failing
in the situation where we're running a local analysis with no
AWS_STORAGE_BUCKET_NAME set.