Upgrade Django and psycopg2 to match base container
The base container has gotten some patch version upgrades and dependabot
is saying we should upgrade Django. Staying in sync with the base container
isn't strictly necessary, but seems like a good plan. So this does that,
and adjusts the comment in the requirements file to better describe the
situation.