Load default speed limit when importing analysis results
Changes the 'upload_local_analysis' task to load the default residential
speed limit from the file where it gets exported by the analysis.
Unlike the other files that are part of the export, it won't crash if
this one is missing, so that exports of older analysis runs will still
be importable. The front-end only shows the speed limit box if there's
a default speed limit set on the job, so leaving it blank when we don't
have the answer is fine.