Fix pip installation and remove azavea.pip ansible role (#828)
The URL of the 'get-pip.py' script now has a version number, which
broke provisioning.
In two places, actually, since we need pip at the very beginning to
install ansible inside the VM, then we also had the Azavea ansible.pip
role being used by the Azavea aws-cli role.

Having a second place where we install pip, especially one that's very
simple but not very actively maintained, isn't worth it. So this clones
the aws-cli role into a local role and drops the dependency on the pip
role. Since pip is installed in the Vagrantfile provisioning, it will
always be there.

Also, Ansible still wants to use Python 2 by default, but that's no good.
This makes Ansible and the AWS CLI both get installed in Python 3.