Comment out Tilegarden deploy step for now
The Tilegarden package has gotten too big. We haven't changed anything,
so it has to be that the dependencies grew. But we need to figure out
how to shrink them again to get it to deploy.
Fortunately we can just keep the old one up until we've done that.
So this comments out the command in 'infra' where we try to redeploy it.