Update AWS CLI to v2 to fix deprecated Python error
We've been installing AWS CLI v1 on the analysis container using the
"download the bundle and run the install script" method, but that no longer
supports older versions of Python.

Rather than have to upgrade the version of Python we're running in the
container, we can just switch to AWS CLI v2, which has no external Python
dependency.  We only use the AWS CLI for simple S3 uploading and
downloading, so moving to v2 doesn't require any changes to the commands
we're running.

I also changed `python` to `python3` in the list of apt-get packages to
install. It was already getting pulled in by `python3-dev`, but I don't
think there's any reason to keep the Python 2 package around, and if there's
a place that we're accidentally still using Python 2, I'd rather find out
about it than let it keep quietly working.