Update base VM to fix ansible installation error (#856)
VM provisioning has started failing, with a message about Python 3.5 having
reached end-of-life. There might be other ways to get around it, but upgrading
the base VM would be a good thing anyway, so that's the first thing I tried.
It seems like it worked--I'm able to provision the app and run the analysis.

I also updated the Ansible roles for good measure. We only use the VM as a
standard environment in which to build and run the containers, so it's not
that surprising that it wouldn't be too susceptible to issues from upgrading.

The new VM did have a problem with the old version of Docker, so I updated
the version in the group_vars/all.example to one that works with the new
base box. I also upgraded Terraform to the last patch version of the minor
version we're using.