Tweak Vagrantfile settings for Virtualbox 6 issues
This started crashing for me after my Virtualbox got upgraded to v6.
There's no reason we need a particular internal IP for this project, so
I just switched it to one in the approved range.
I'm also not sure why the vers=3 option for NFS was needed before, or
why it's a problem now, but I just checked what I was doing in Cicero
(not specifying it either way) and went with that.