Make max_trip_distance optional in AnalysisBatch.create_from_shapefile
Adjusts the `create_from_shapefile` method on AnalysisBatch to make the
max_trip_distance param optional. Specifically, the argument defaults to
None but when the value is None it omits the argument when it calls
AnalysisJob.objects.create so that the model default will apply.

Since we already have working create_from_shapefile tests to copy from,
this also adds a couple tests to confirm that the default and override
cases work.